moved functions to be independent of CRUD
